requirements.txt
cuesdk==3.1.1
Pillow==9.2.0